основной код

+++ stats-service/stats-dto/src/main/java/ru/practicum/ewm/dto/StatsResponseDto.java
@@ -8,5 +8,5 @@ import lombok.Data;
 public class StatsResponseDto {
     private String app;
     private String uri;
-    private Integer hits;
+    private Long hits;
 }
