+++ stats-service/stats-server/src/main/java/ru/practicum/ewm/mapper/StatsResponseDtoRowMapper.java
@@ -14,7 +14,7 @@ public class StatsResponseDtoRowMapper implements RowMapper<StatsResponseDto> {
         return StatsResponseDto.builder()
                 .app(rs.getString("app"))
                 .uri(rs.getString("uri"))
-                .hits(rs.getInt("hits"))
+                .hits(rs.getLong("hits"))
                 .build();
     }
 }
